% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{add_to_slide}
\alias{add_to_slide}
\title{Add item to slide.  slide and item must both be list items}
\usage{
add_to_slide(slide, item)
}
\arguments{
\item{slide}{A slide to add the item to.}

\item{item}{An item to add to the slide. E.g. output of text_box(), new_custom_shape().}
}
\value{
The input slide with the new item added to it.
}
\description{
Update a slide list item to add a new item to it. In practice this will
likely be text_box() or new_custom_shape() objects.
}
