% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDistmat.R
\name{CheckDistmat}
\alias{CheckDistmat}
\title{Check distance matrices}
\usage{
CheckDistmat(distmat)
}
\arguments{
\item{distmat}{An \eqn{n}-by-\eqn{n} matrix holding the pairwise distances between observations.}
}
\description{
Check whether the input distance matrix or its sub square matrix is symmetric 
and has zero diagonals.
}
