% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{create_metadata}
\alias{create_metadata}
\title{Create a metadata list object.}
\usage{
create_metadata(
  title = NA,
  description = NA,
  contributor = NA,
  keyword = NA,
  coverage = NA,
  identifier = NA,
  publisher = NA,
  relation = NA,
  rights = NA,
  source = NA,
  type = NA,
  subject = NA
)
}
\arguments{
\item{title}{Character. Title of presentation.}

\item{description}{Character.}

\item{contributor}{Character.}

\item{keyword}{Character.}

\item{coverage}{Character.}

\item{identifier}{Character.}

\item{publisher}{Character.}

\item{relation}{Character.}

\item{rights}{Character.}

\item{source}{Character.}

\item{type}{Character.}

\item{subject}{Character.}
}
\description{
Must be written to disk with write_metadata().
}
