% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{draw_enhanced_geometry_list}
\alias{draw_enhanced_geometry_list}
\title{Create enhanced geometry item. Not used by user}
\usage{
draw_enhanced_geometry_list(type = c("ellipse", "rectangle"))
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
