% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{draw_frame_list}
\alias{draw_frame_list}
\title{Create a \code{draw:frame} list object.}
\usage{
draw_frame_list(
  width,
  height,
  x,
  y,
  draw_layer = "layout",
  draw_style_name = "gr1",
  draw_text_style_name = "P1"
)
}
\arguments{
\item{width}{Character. Width in cm. e.g. "10cm"}

\item{height}{Character. Height in cm. e.g. "10cm"}

\item{x}{Character. Leftmost position in cm. e.g. "10cm"}

\item{y}{Character. Topmost position in cm. e.g. "10cm"}

\item{draw_layer}{Character. Default "layout".}

\item{draw_style_name}{Character. The draw style to apply. Default "gr1".}

\item{draw_text_style_name}{Character. The text style to apply. Default "P1".}
}
\value{
A draw:frame list object.
}
\description{
Description of what the function does.
how duplication is measured.
}
