% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Expit function}
\usage{
expit(x)
}
\arguments{
\item{x}{A vector holding the real valued input.}
}
\value{
A vector holding the result.
}
\description{
Compute the value of \eqn{\exp(x)/(1 + \exp(x))} for given \eqn{x}.
}
