% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{field_page_num_list}
\alias{field_page_num_list}
\title{A field that will display the current page number. goes inside a text box}
\usage{
field_page_num_list()
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
