% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDistPrfl.R
\name{GetDistPrfl}
\alias{GetDistPrfl}
\title{Compute distance profiles with respect to a sample of random objects}
\usage{
GetDistPrfl(
  distmat,
  data,
  distfun = NULL,
  newdata = NULL,
  type = "qf",
  optns = list(),
  ...
)
}
\arguments{
\item{distmat}{An \eqn{(n+m)}-by-\eqn{n} matrix holding the pairwise distances between observations 
in training data of size \eqn{n} and possibly new data of size \eqn{m} to the training data.
If there is no new data of interest, then \code{distmat} should be a symmetric matrix of dimension \eqn{n}. 
At least one of \code{data} and \code{distmat} should be input. 
If both are given, \code{distmat} overrides \code{data}.}

\item{data}{Training data with respect to the law of which distance profiles are of interest. 
Either a matrix or data frame with \eqn{n} rows where each row holds one observation 
in the training data, or a list (but not a data frame) of length \eqn{n} where each element holds one observation.}

\item{distfun}{A function with two arguments computing the distance between two observations, 
which is used in \code{\link{GetPairDist}}. Default: Euclidean distance.}

\item{newdata}{Optional new data, at which distance profiles are to be computed 
with respect to the law of \code{data}. 
Either a matrix or data frame with \eqn{m} rows and the same number of columns as \code{data} 
where each row holds one new observation in addition to the observations in \code{data}. 
or a list (but not a data frame) of length \eqn{m} where each element is of the same format as those in \code{data} 
and holds one new observation. 
This can only be specified if \code{data} but not \code{distmat} is given.}

\item{type}{Character specifying the type of results to be computed:
\code{'den'} for densities,
\code{'cdf'} for cumulative distribution functions,
\code{'qf'} (default) for quantile functions,
and \code{'all'} for all of the three aforementioned types.}

\item{optns}{A list of options control parameters specified by \code{list(name=value)}. See 'Details'.}

\item{...}{Optional arguments of \code{distfun}.}
}
\value{
A list. All three fields \code{den} (for the densities), \code{cdf} (for the cdfs) 
and \code{qf} (for the quantile functions) will be included if \code{type == 'all'}, 
and the corresponding field alone out of the three otherwise.
\item{den}{A list of \eqn{n} or \eqn{(n+m)} fields if \code{optns$outputGrid} is specified, 
each of which is a list of three fields, \code{bw}, \code{x} and \code{y}; 
see 'Value' of \code{\link[frechet]{CreateDensity}} for further details.
If \code{optns$outputGrid} is unspecified, 
it is a list of three fields, \code{bwvec}, \code{x} and \code{ymat}, where 
\code{bwvec} is a vector of length \eqn{n} or \eqn{(n+m)} holding the bandwidths used for smoothing,
\code{x} is the (common) support grid of the \eqn{n} or \eqn{(n+m)} densities specified by \code{optns$outputGrid}, 
and \code{ymat} is a matrix with \eqn{n} or \eqn{(n+m)} rows, each row holding the values of the density for one subject.}
\item{cdf}{A list of \eqn{n} or \eqn{(n+m)} fields if \code{optns$outputGrid} is specified, 
each of which is a list of two fields, \code{x} and \code{y}, 
which are two vectors holding the support grid and the corresponding values of the cdf, respectively.
If \code{optns$outputGrid} is unspecified, 
it is a list of two fields, \code{x} and \code{ymat}, where 
\code{x} is the (common) support grid of the \eqn{n} or \eqn{(n+m)} cdfs specified by \code{optns$outputGrid}, 
and \code{ymat} is a matrix with \eqn{n} or \eqn{(n+m)} rows, each row holding the values of the cdf for one subject.}
\item{qf}{A list of two fields, \code{x} and \code{ymat}, where 
\code{x} is a vector holding the (common) support grid of the \eqn{n} or \eqn{(n+m)} quantile functions, 
and \code{ymat} is a matrix with \eqn{n} or \eqn{(n+m)} rows, each row holding the values of the quantile function for one subject.}
\item{n}{The sample size of the training data.}
\item{distmat}{An \eqn{(n+m)}-by-\eqn{n} matrix holding the pairwise distances between observations 
in training data of size \eqn{n} and possibly new data of size \eqn{m} to the training data. 
This will be output only if \code{data} but not \code{distmat} is given in the input.}
}
\description{
Compute the distribution of \eqn{d(x,X)} for given \eqn{x}. 
For \eqn{x = X_i} within the training data \eqn{\{X_i\}_{i=1}^{n}}, 
the distribution of \eqn{d(x,X)} is estimated based on \eqn{\{d(x,X_j)\}_{j\neq i}}.
For a new \eqn{x} out of the training data, 
the distribution of \eqn{d(x,X)} is estimated based on \eqn{\{d(x,X_j)\}_{j=1}^{n}}.
}
\details{
Cumulative distribution function and quantile function representations of 
distance profiles are obtained as the right-continuous empirical distribution functions 
and their left-continuous inverse, respectively. 
If density representation of distance profiles is of interest 
(\code{type = 'den'} or \code{type = 'all'}), 
density estimation is performed using \code{\link[frechet]{CreateDensity}}.
Available control options are \code{userBwMu}, \code{nRegGrid}, \code{delta}, \code{kernel}, 
\code{infSupport}, \code{outputGrid}, \code{nqSup} and \code{qSup}. 
See \code{\link[frechet]{CreateDensity}} for the details about the options 
other than \code{nqSup} and \code{qSup}.
\describe{
\item{nqSup}{The number of equidistant points in the support grid of the quantile functions. 
Default: 101.}
\item{qSup}{The support grid of the quantile functions; it overrides \code{nqSup}. 
Default: \code{seq(0,1,length.out=optns$nqSup)}.}
}
}
\examples{
d <- 2
n <- 10
m <- 5
data <- matrix( rnorm( n * d ), ncol = d )
newdata <- matrix( rnorm( m * d ), ncol = d )
distmat <- as.matrix( stats::dist( rbind(data,newdata) ) )[, 1:n, drop=FALSE]
## with input distmat
res <- GetDistPrfl( distmat = distmat )
## Or with input data and newdata
res2 <- GetDistPrfl( data = data, newdata = newdata )
}
