% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPrflHomTestStat.R
\name{GetPrflHomTestStat}
\alias{GetPrflHomTestStat}
\title{Compute the test statistic of the distance profile based two-sample homogeneity test for random objects}
\usage{
GetPrflHomTestStat(distmat, n, nRegGrid = 101)
}
\arguments{
\item{distmat}{An \eqn{(n+m)}-by-\eqn{(n+m)} matrix holding the pairwise distances between observations 
in the two samples of sizes \eqn{n} and \eqn{m}, respectively.}

\item{n}{Size of the first sample.}

\item{nRegGrid}{Number of equidistant grid points from the minimum to the maximum of all pairwise distances,
on which the integrals of difference between distance profiles will be taken. Default: 101.}
}
\value{
The value of the test statistic.
}
\description{
Compute the test statistic of the two-sample test for random objects based on comparing the distance profiles.
}
\examples{
n <- m <- 50
data <- rep( rnorm(n,0,1), 2 )
distmat <- abs( matrix( data, nr = n+m, nc = n+m ) - matrix( data, nr = n+m, nc = n+m, byrow = TRUE ) )
stat <- GetPrflHomTestStat( distmat = distmat, n = n )
}
