% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{image_list}
\alias{image_list}
\title{Create an image item in list format}
\usage{
image_list(
  img_filepath,
  width,
  height,
  x,
  y,
  alt_text,
  draw_layer = "layout",
  draw_style_name = "gr2",
  draw_text_style_name = "P1"
)
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
