% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2metric.R
\name{l2metric}
\alias{l2metric}
\title{\eqn{L^2} metric between square integrable functions.}
\usage{
l2metric(x, y, sup)
}
\arguments{
\item{x, y}{Vectors holding the values of two functions evaluated on \code{sup}.}

\item{sup}{Support grid.}
}
\value{
The \eqn{L^2} metric between the two functions.
}
\description{
\eqn{L^2} metric between square integrable functions.
}
