% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{list_item_to_xml}
\alias{list_item_to_xml}
\title{Recursive funciton to parse list into xml???}
\usage{
list_item_to_xml(item)
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
