% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeObjMdsPlot.R
\name{MakeObjMdsPlot}
\alias{MakeObjMdsPlot}
\title{Make an object MDS plot}
\usage{
MakeObjMdsPlot(
  distmat,
  data,
  color_by,
  nGroup,
  shape_by = NULL,
  shape = 3,
  id,
  id_size = 3,
  xlab = "Coordinate 1",
  ylab = "Coordinate 2",
  colorlab = NULL,
  shapelab = NULL,
  use_default_colors = TRUE
)
}
\arguments{
\item{distmat}{An \eqn{n}-by-\eqn{n} symmetric matrix holding the pairwise distances between observations.}

\item{data}{An \eqn{n}-by-2 matrix or data frame holding the data; 
This is only applicable for 2-dimensional Euclidean data and overrides \code{distmat} if there is a proper input.}

\item{color_by}{A vector of length \eqn{n} holding the variable according to which 
colors are assigned to each point. If missing, all points will be in black.}

\item{nGroup}{Number of groups for optional grouping according to \code{color_by}. 
If given, the sample is divided into \code{nGroup} groups according to quantiles of \code{color_by}, 
which should be numeric in this case; the \eqn{i}-th group contains the 
points with the corresponding value in \code{color_by} falling in 
\eqn{( q_{1-i/\code{nGroup}}, q_{1-(i-1)/\code{nGroup}} ] }, 
except for \eqn{i=\code{nGroup}}, for which the corresponding interval is \eqn{[ q_{0}, q_{1/\code{nGroup}} ]}. 
Here, \eqn{q_{\alpha}} is the \eqn{\alpha}-quantile of \code{color_by}, given by \code{quantile(color_by,alpha)}.}

\item{shape_by}{A vector of length \eqn{n} holding the variable according to which 
shapes are assigned to each point. Default: \code{NULL}.}

\item{shape}{Shape of points when \code{shape_by} is \code{NULL}. Default: 3. 
See \code{\link[ggplot2]{geom_point}} for details.}

\item{id}{A vector of length \eqn{n} holding the label of each point. 
Default: \code{rownames(data)} if \code{data} is given and 
\code{rownames(distmat)} otherwise.}

\item{id_size}{Size of the text labels of points. Default: 3.}

\item{xlab, ylab}{The labels of x-axis and y-axis. Default: \code{'Coordinate 1'} and \code{'Coordinate 2'}.}

\item{colorlab}{The label of the color variable when \code{color_by} is given. Default: \code{NULL}.}

\item{shapelab}{The label of the shape variable when \code{shape_by} is given. Default: \code{NULL}.}

\item{use_default_colors}{Logical, whether to use default colors when \code{color_by} is given. Default: \code{TRUE}.}
}
\value{
A ggplot object.
}
\description{
Make a plot of 2-dimensional classical metric multidimensional scaling of the 
data based on the pairwise distance matrix or a scatterplot of 2-dimensional Euclidean data.
}
\examples{
n <- 100
p <- 2
set.seed(1)
data <- matrix( rnorm( n * p ), ncol = p )
res <- GetTpRank( data = data )
pl <- MakeObjMdsPlot( data = data, color_by = res$rank, nGroup = 10 )
}
\seealso{
\code{\link{GetTpRank}}, \code{\link{MakePrflMdsPlot}}
}
