% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_font}
\alias{new_font}
\title{Define new font for use in the deck.}
\usage{
new_font(name, font_family_generic = "system", font_pitch = "variable")
}
\arguments{
\item{name}{Character. Name of the font.}

\item{font_family_generic}{Character. Default "system".}

\item{font_pitch}{Character. Default "variable".}
}
\value{
A new font declaration in list format.
}
\description{
Does not install fonts or check that they are available.
}
