% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_graphic_style_list}
\alias{new_graphic_style_list}
\title{Define a new graphics style.}
\usage{
new_graphic_style_list(
  name,
  stroke_color = "#000000",
  stroke_opacity = "100\%",
  fill_type = c("none", "solid"),
  fill_color = "#FFFFFF",
  fill_opacity = "100\%",
  decorative = TRUE
)
}
\arguments{
\item{name}{Character. Name of the style, used later to apply it.}

\item{stroke_color}{Character. Stroke colour in hex format. Default "#000000".}

\item{stroke_opacity}{Character. Default "100\%".}

\item{fill_type}{Character. c("none", "solid").}

\item{fill_color}{Character. Default "#FFFFFF".}

\item{fill_opacity}{Character. Default "100\%".}

\item{decorative}{Boolean. Default TRUE.}
}
\value{
A new graphic style item.
}
\description{
Description of what the function does.
how duplication is measured.
}
