% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_list}
\alias{new_list}
\title{Create a list}
\usage{
new_list(list_contents, text_style_name = NA)
}
\arguments{
\item{list_contents}{Character vector or list. Character vector will give
an unordered list. List will give a sublist.}

\item{text_style_name}{Character. name of text style.}
}
\description{
Must be embedded in a text_box().
}
