% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_paragraph_style}
\alias{new_paragraph_style}
\title{Create new paragraph style.}
\usage{
new_paragraph_style(
  name,
  font_weight = c("regular", "bold"),
  font_style = c("regular", "italic"),
  font_size = "12pt",
  color = "#000000",
  text_align = c("start", "center", "end"),
  opacity = "100\%",
  font_name = "Liberation Sans",
  text_underline_style = c("none", "solid")
)
}
\arguments{
\item{name}{The style's name. This is used in text_box() to apply the style.}

\item{font_weight}{Character. c("regular", "bold").}

\item{font_style}{Character. c("regular", "italic").}

\item{font_size}{Character. Font size in pts. Default "12pt".}

\item{color}{Character. Text colour in hex format. Default "#000000".}

\item{text_align}{Character. c("start", "center", "end").}

\item{opacity}{Character. Opacity in percent. Default "100\%".}

\item{font_name}{character. Default "Liberation Sans".}

\item{text_underline_style}{Character. c("none", "solid")).}
}
\value{
A paragraph style list item.
}
\description{
Description of what the function does.
how duplication is measured.
}
