% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_paragraph_style_list}
\alias{new_paragraph_style_list}
\title{Create new paragraph style.}
\usage{
new_paragraph_style_list(
  name,
  font_weight = c("regular", "bold"),
  font_style = c("regular", "italic"),
  font_size = "12pt",
  color = "#000000",
  text_align = c("start", "center", "end"),
  opacity = "100\%",
  font_name = "Liberation Sans"
)
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
