% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_pres}
\alias{new_pres}
\title{Create a new presentation. This returns an XML object and also has side effects on disk.}
\usage{
new_pres()
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
