% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_text_style}
\alias{new_text_style}
\title{Define new text style.}
\usage{
new_text_style(
  name,
  font_weight = c("regular", "bold"),
  font_style = c("regular", "italic"),
  font_size = "12pt",
  color = "#000000",
  opacity = "100\%",
  font_name = "Liberation Sans",
  text_underline_style = c("none", "solid")
)
}
\arguments{
\item{name}{Character. Name of style that will be used to apply it.}

\item{font_weight}{Character. Default  c("regular", "bold")}

\item{font_style}{Character. Default  c("regular", "italic")}

\item{font_size}{Character. Default  "12pt"}

\item{color}{Character. Default  "#000000"}

\item{opacity}{Character. Default  "100\%"}

\item{font_name}{Character. Default  "Liberation Sans"}

\item{text_underline_style}{Character. Default  c("none", "solid"))}
}
\value{
A new text style.
}
\description{
Does overwrite other style elements.
}
