% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_text_style_minimal}
\alias{new_text_style_minimal}
\title{Define new minimal text style.}
\usage{
new_text_style_minimal(
  name,
  font_weight = NA,
  font_style = NA,
  font_size = NA,
  color = NA,
  opacity = NA,
  font_name = NA,
  text_underline_style = NA,
  text_underline_width = NA,
  text_underline_color = NA
)
}
\value{
A new minimal text style.
}
\description{
Does not overwrite other style elements.
}
