% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{save_pres}
\alias{save_pres}
\title{Save presentation as compressed ODP file in the current working directory and return the input document.}
\usage{
save_pres(doc, filename)
}
\arguments{
\item{doc}{The deck object to save.}

\item{filename}{The filename in the current working folder to save the output.}
}
\value{
The input deck object, invisibly.
}
\description{
This function receives a final deck object, performs disk operations as a side effect to create
an output file, and then returns the original deck object invisibly.
}
\details{
This is the \emph{final step} and should be done after calling write_fonts(), write_styles(), write_manifest(),
and write_slides(). They can be conveniently piped together, with \code{save_pres()} as the final step.
}
