% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{text_box}
\alias{text_box}
\title{Create a text box list including its wrapping draw_frame}
\usage{
text_box(
  text,
  width,
  height,
  x,
  y,
  draw_layer = "layout",
  draw_style_name = "gr1",
  draw_text_style_name = "P1"
)
}
\arguments{
\item{text}{List. The text to display as a series of text_p() or new_list() objects.}

\item{width}{Character. Width in cm. e.g. "10cm"}

\item{height}{Character. Height in cm. e.g. "10cm"}

\item{x}{Character. Leftmost position in cm. e.g. "10cm"}

\item{y}{Character. Topmost position in cm. e.g. "10cm"}

\item{draw_layer}{Character. Default "layout".}

\item{draw_style_name}{Character. The draw style to apply. Default "gr1".}

\item{draw_text_style_name}{Character. The text style to apply. Default "P1".}
}
\value{
A textbox list item.
}
\description{
This is the main function for adding text to your slides. Text boxes are
the highest-level text objects, and they set the geometry for the text:
where it will go, how large the shape will be, and so on.
}
\details{
Text inside the boxes is provided as a list of text_p() (paragraph) or
new_list() (list) objects.
}
