% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{text_box2}
\alias{text_box2}
\title{Internal function that just creates the  text box, but the main funciton also wraps it in a draw:frame}
\usage{
text_box2(text, draw_text_style_name)
}
\arguments{
\item{text}{Character. The text to display.}

\item{draw_text_style_name}{Character. The name of the text style to apply.}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
