% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{text_p_list}
\alias{text_p_list}
\title{Create a text:p item in list format}
\usage{
text_p_list(text, text_style_name)
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
