% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{text_span}
\alias{text_span}
\title{Create a text span in list format}
\usage{
text_span(text, style_name = NA)
}
\arguments{
\item{text}{Character or field list object.}

\item{style_name}{Character. Optional. Name of new_text_style() or new_text_style_minimal() object.}
}
\value{
A text_span list object.
}
\description{
Spans are the smallest unit of text. They can be individually styled with a new_text_style()
or new_text_style_minimal() style object. They must live inside of text_p() objects. They do
not create linebreaks or whitespace if they are next to each other.
}
\details{
Spans can contain either raw character text or field objects, e.g. from field_page_num().
}
