% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{write_manifest}
\alias{write_manifest}
\title{Write the manifest.xml for any images added to the presentation}
\usage{
write_manifest(deck)
}
\arguments{
\item{deck}{A deck object.}
}
\value{
The deck unmodified, but has side-effects on disk.
}
\description{
Returns the deck object unmodified, but has side-effects on disk.
}
