% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{write_metadata}
\alias{write_metadata}
\title{Write metadata (title, subject, etc.) to disk}
\usage{
write_metadata(doc, metadata)
}
\arguments{
\item{doc}{A deck object created with new_pres().}

\item{metadata}{A metadata list object created with create_metadata().contents_list List. A list of text_p() or new_list() items prepared previously.}
}
\description{
Write metadata (title, subject, etc.) to disk
}
