% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{write_slides}
\alias{write_slides}
\title{Write slides in deck object. Slides must be list of list style items}
\usage{
write_slides(doc, slides)
}
\arguments{
\item{doc}{A deck object.}

\item{slides}{A list of slides.}
}
\value{
The deck object with the slides added.
}
\description{
Update the deck object to write user-defined slides. This should only be
run once in a final pipeline that ends with \code{save_pres()}.
}
