% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{write_styles}
\alias{write_styles}
\title{Write styles in deck object. Styles must be list of list style items}
\usage{
write_styles(doc, styles)
}
\arguments{
\item{doc}{A deck object.}

\item{styles}{A list of styles created by a style function, e.g. new_paragraph_style() or
odp::new_graphics_style().}
}
\value{
An updated deck object with the styles added to it.
}
\description{
Update the deck object with user-defined styles.
}
