% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{add_to_slide}
\alias{add_to_slide}
\title{Add item to slide.  slide and item must both be list items}
\usage{
add_to_slide(slide, item)
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
