% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{add_to_slide}
\alias{add_to_slide}
\title{Add item to slide.  slide and item must both be list items}
\usage{
add_to_slide(slide, item)
}
\arguments{
\item{slide}{A slide to add the item to.}

\item{item}{An item to add to the slide. E.g. output of text_box_list(), new_custom_shape_list().}
}
\value{
The input slide with the new item added to it.
}
\description{
Description of what the function does.
how duplication is measured.
}
