% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{draw_frame_list}
\alias{draw_frame_list}
\title{Create a \code{draw:frame} list object.}
\usage{
draw_frame_list(
  width,
  height,
  x,
  y,
  draw_layer = "layout",
  draw_style_name = "gr1",
  draw_text_style_name = "P1"
)
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
