% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{field_page_num_list}
\alias{field_page_num_list}
\title{A field for the current page/slide number.}
\usage{
field_page_num_list()
}
\value{
A field showing the current page number.
}
\description{
This field item should be used inside a text box.
}
