% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{image_list}
\alias{image_list}
\title{Create an image item in list format}
\usage{
image_list(
  img_filepath,
  width,
  height,
  x,
  y,
  alt_text,
  draw_layer = "layout",
  draw_style_name = "gr1",
  draw_text_style_name = "P1"
)
}
\arguments{
\item{img_filepath}{Character. Path to image to be inserted.}

\item{width}{Character. Width in cm. e.g. "10cm"}

\item{height}{Character. Height in cm. e.g. "10cm"}

\item{x}{Character. Leftmost position in cm. e.g. "10cm"}

\item{y}{Character. Topmost position in cm. e.g. "10cm"}

\item{alt_text}{Character. Alt text for image. Important for accessibility.}

\item{draw_layer}{Character. Default "layout".}

\item{draw_style_name}{Character. The draw style to apply. Default "gr1".}

\item{draw_text_style_name}{Character. The text style to apply. Default "P1".}
}
\value{
An image object that can be added to a slide.
}
\description{
Only works with png or svg images. This function has side
effects and creates a temp copy of the image on disk to it
can be embedded in the output odp file.
}
