% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{list_item_to_xml}
\alias{list_item_to_xml}
\title{Recursive funciton to parse list into xml.}
\usage{
list_item_to_xml(item)
}
\arguments{
\item{item}{A list representing an XML data structure. Each node is
represented as a list with the following named members.
type: Character. The type of node. E.g. "draw:image"
attrbutes: Named character vector, with names as attribute
names and values as attribute values.
children:  NULL or character or list. If character, raw text.
If list, list of child nodes of this same type.}
}
\value{
An XML tree created by the package xml2.
}
\description{
Recursive funciton to parse list into xml.
}
