% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_custom_shape}
\alias{new_custom_shape}
\title{Create new custom shape.}
\usage{
new_custom_shape(
  type = c("rectangle", "ellipse", "round-rectangle"),
  width,
  height,
  x,
  y,
  draw_style_name = "gr1",
  text_style_name = "P1",
  text = "",
  rect_radius = 0
)
}
\arguments{
\item{type}{Character. c("rectangle", "ellipse")}

\item{width}{Character. Width in cm. e.g. "10cm"}

\item{height}{Character. Height in cm. e.g. "10cm"}

\item{x}{Character. Leftmost position in cm. e.g. "10cm"}

\item{y}{Character. Topmost position in cm. e.g. "10cm"}

\item{draw_style_name}{Character. The draw style to apply. Default "gr1".}

\item{text_style_name}{Character. The text style to apply. Default "P1".}

\item{text}{Character. Text to include in shape. Default "".}

\item{rect_radius}{Numeric. Value between 0 and 10800, higher numbers give rounder rounded rectangles.}
}
\value{
A custom shape list object.
}
\description{
Description of what the function does.
how duplication is measured.
}
