% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_custom_shape_list}
\alias{new_custom_shape_list}
\title{Create new custom shape.}
\usage{
new_custom_shape_list(
  type = c("rectangle", "ellipse"),
  width,
  height,
  x,
  y,
  draw_style_name = "gr1",
  text_style_name = "P1",
  text = ""
)
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
