% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_font_list}
\alias{new_font_list}
\title{Define new font for use in the deck.}
\usage{
new_font_list(name, font_family_generic = "system", font_pitch = "variable")
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
