% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_graphic_style}
\alias{new_graphic_style}
\title{Define a new graphics style.}
\usage{
new_graphic_style(
  name,
  stroke_color = "#000000",
  stroke_opacity = "100\%",
  fill_type = c("none", "solid"),
  fill_color = "#FFFFFF",
  fill_opacity = "100\%",
  decorative = TRUE
)
}
\arguments{
\item{name}{Character. Name of the style, used later to apply it.}

\item{stroke_color}{Character. Stroke colour in hex format. Default "#000000".}

\item{stroke_opacity}{Character. Default "100\%".}

\item{fill_type}{Character. c("none", "solid").}

\item{fill_color}{Character. Default "#FFFFFF".}

\item{fill_opacity}{Character. Default "100\%".}

\item{decorative}{Boolean. Default TRUE.}
}
\value{
A new graphic style item.
}
\description{
Create a new graphics style with sensible defaults.
}
\details{
To use this style you need to do two things. First, make sure that all
styles are put into a list and saved as part of your final project pipeline
with the function \code{write_styles()}.

Second, ensure that you use each style's name as defined in its \code{name} attribute
when you created it. This is similar to CSS, where a style is defined and named
and then referred back to at the point of use.

Graphics styles can be used in new_custom_shape() function calls.
}
