% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_graphic_style_list}
\alias{new_graphic_style_list}
\title{Define a new graphics style.}
\usage{
new_graphic_style_list(
  name,
  stroke_color = "#000000",
  stroke_opacity = "100\%",
  fill_type = c("none", "solid"),
  fill_color = "#FFFFFF",
  fill_opacity = "100\%",
  decorative = TRUE
)
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
