% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_pres}
\alias{new_pres}
\title{Create a new presentation.}
\usage{
new_pres()
}
\value{
A new presentation object, plus side effects on disk to set it up.
}
\description{
This returns an XML object and also has side effects on disk.
}
