% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_slide}
\alias{new_slide}
\title{Create a slide item in list format}
\usage{
new_slide(name = "slide")
}
\arguments{
\item{name}{Character. Slide name. Will go in e.g. pdf index. Important for accessibility.}
}
\value{
A slide object in list format.
}
\description{
Each slide is a new page in your presentation. Slides are defined as lists. Objects are added
to slides with the add_to_slide() function.
}
