% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_text_style}
\alias{new_text_style}
\title{Define new text style.}
\usage{
new_text_style(
  name,
  font_weight = c("regular", "bold"),
  font_style = c("regular", "italic"),
  font_size = "12pt",
  color = "#000000",
  opacity = "100\%",
  font_name = "Liberation Sans",
  text_underline_style = c("none", "solid")
)
}
\value{
A new text style.
}
\description{
Does overwrite other style elements.
}
