% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{new_text_style_minimal}
\alias{new_text_style_minimal}
\title{Define new minimal text style.}
\usage{
new_text_style_minimal(
  name,
  font_weight = NA,
  font_style = NA,
  font_size = NA,
  color = NA,
  opacity = NA,
  font_name = NA,
  text_underline_style = NA,
  text_underline_width = NA,
  text_underline_color = NA
)
}
\arguments{
\item{name}{Character. Name that will be used to apply the style.}

\item{font_weight}{Character. Default NA}

\item{font_style}{Character. Default NA}

\item{font_size}{Character. Default NA}

\item{color}{Character. Default NA}

\item{opacity}{Character. Default NA}

\item{font_name}{Character. Default NA}

\item{text_underline_style}{Character. Default NA}

\item{text_underline_width}{Character. Default NA}

\item{text_underline_color}{Character. Default NA}
}
\value{
A new minimal text style.
}
\description{
Does not overwrite other style elements; has no defaults. For options,
see new_text_style() documentation.
}
