% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{save_pres}
\alias{save_pres}
\title{Save presentation as compressed ODP file in the current working directory and return the input document.}
\usage{
save_pres(doc, filename)
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
