% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{slide_list}
\alias{slide_list}
\title{Create a slide item in list format}
\usage{
slide_list(name = "slide")
}
\arguments{
\item{name}{Character. Slide name. Will go in e.g. pdf index. Important for accessibility.}
}
\value{
Description of what the function returns.
}
\description{
Create a slide item in list format
}
