% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{text_box_list}
\alias{text_box_list}
\title{Create a textbox list including its wrapping draw_frame}
\usage{
text_box_list(
  text,
  width,
  height,
  x,
  y,
  draw_layer = "layout",
  draw_style_name = "gr1",
  draw_text_style_name = "P1"
)
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
