% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{text_box_list2}
\alias{text_box_list2}
\title{Function that just creates the  text box, but the main funciton also wraps it in a draw:frame}
\usage{
text_box_list2(text, draw_text_style_name)
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
