% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{text_p}
\alias{text_p}
\title{Create a text:p item in list format}
\usage{
text_p(text, text_style_name)
}
\arguments{
\item{text}{Character. The text to display.}

\item{text_style_name}{Character. The name of the text style to apply.}
}
\value{
A list of one or more text:p list items.
}
\description{
Description of what the function does.
how duplication is measured.
}
