% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{write_fonts}
\alias{write_fonts}
\title{Write fonts in deck object. Fonts must be list of list font items}
\usage{
write_fonts(doc, fonts)
}
\arguments{
\item{doc}{A deck object.}

\item{fonts}{A list of font declarations from new_font().}
}
\value{
The deck object with the fonts applied.
}
\description{
Update deck object to write user-defined fonts. Does not include fonts in
the file; users must have fonts installed for this to work properly.
}
\details{
Should be run once in a final pipeline terminating in \code{save_pres()}.
}
