% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{write_fonts}
\alias{write_fonts}
\title{Write fonts in deck object. Fonts must be list of list font items}
\usage{
write_fonts(doc, fonts)
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
