% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{write_fonts}
\alias{write_fonts}
\title{Write fonts in deck object. Fonts must be list of list font items}
\usage{
write_fonts(doc, fonts)
}
\arguments{
\item{doc}{A deck object.}

\item{fonts}{A list of font declarations from new_font_list().}
}
\value{
The deck object with the fonts applied.
}
\description{
Description of what the function does.
how duplication is measured.
}
