% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{write_slides}
\alias{write_slides}
\title{Write slides in deck object. Slides must be list of list style items}
\usage{
write_slides(doc, slides)
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
