% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{write_styles}
\alias{write_styles}
\title{Write styles in deck object. Styles must be list of list style items}
\usage{
write_styles(doc, styles)
}
\arguments{
\item{doc}{A deck object.}

\item{styles}{A list of styles created new_paragraph_style_list() or
odp::new_graphics_style_list().}
}
\value{
An updated deck object with the styles added to it.
}
\description{
Description of what the function does.
how duplication is measured.
}
