% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odt_functions.R
\name{write_styles}
\alias{write_styles}
\title{Write styles in deck object. Styles must be list of list style items}
\usage{
write_styles(doc, styles)
}
\arguments{
\item{type}{description}
}
\value{
Description of what the function returns.
}
\description{
Description of what the function does.
how duplication is measured.
}
